requirements.txt
langchain
langchain_community
langchain-google-genai
python-dotenv
streamlit
langchain_chroma
pypdf